refactor: LottoController의 run() 에서 실행 흐름을 모두 관리하도록 변경

+++ src/main/java/lotto/domain/User.java
@@ -7,8 +7,8 @@ import lotto.service.LottoSeller;
 
 public class User {
     private final List<Lotto> lottoList;
-    private int purchaseAmount;
-    private int totalEarnings;
+    private long purchaseAmount;
+    private long totalEarnings;
     private double yield;
 
 
@@ -33,15 +33,15 @@ public class User {
         this.purchaseAmount = amount;
     }
 
-    public void setTotalEarnings(int earnings) {
+    public void setTotalEarnings(long earnings) {
         this.totalEarnings = earnings;
     }
 
-    public int getPurchaseAmount() {
+    public long getPurchaseAmount() {
         return purchaseAmount;
     }
 
-    public int getTotalEarnings() {
+    public long getTotalEarnings() {
         return totalEarnings;
     }
